Move the TensorFlow imports into the methods.
Importing the tensorflow packages can take an extraordinarily long
time (minutes on contended IO).  This is annoying for the user if
all they are doing is running the script with --help, or if it's
going to fail.  Not only that but they output a lot of probably
unnecessary and confusing guff to stderr.

By moving the imports into the methods they'll only trigger when
needed.