Address Jim's suggestions (#4)
* Applied black

* Third-party imports after natives

* Added module docstring

* Moved import to toplevel

* Refactoring: python string magic

* Fixing mypy error (?)

* Moved import to top level

* Pythonised print in string

* Added type-hint suggestion

* Added suggested in TODO

* Added type-hinting suggestion in TODO

* Added type-hinting and doc-string suggestions

* Added suggestion about descriptive names as TODO

* Added suggestion in TODO for unused variable

* Updated print statement to use f-string

* Added question in TODO

* Fixed missing word in TODO

* First docstring

* Move the TensorFlow imports into the methods.

Importing the tensorflow packages can take an extraordinarily long
time (minutes on contended IO).  This is annoying for the user if
all they are doing is running the script with --help, or if it's
going to fail.  Not only that but they output a lot of probably
unnecessary and confusing guff to stderr.

By moving the imports into the methods they'll only trigger when
needed.

* Improve the Click help messages.

* Improve the docstrings.

* Add some useful comments

* Add typing information.

Need packages types-click, types-tensorflow.  The latter is
incomplete (of course).

* Linting and style improvements.

Addressing the remainder of Jim's comments.

* Add test for output.

Test that the generated code for 'my_model' has created a function for
associating the appropriate array type and size.

* Fixed docstring formatting.

---------

Co-authored-by: jdenholm <j.denholm.2017@gmail.com>