Fix packaging bug (PackageLoader fails to init)
If you don't include the templates directory in the package
then when you try to initialise the Jinja Loader it fails somewhat
cryptically with
`ValueError: The 'process_model' package was not installed in a way that
PackageLoader understands.`

This does not occur if `pip install -e process_model/` was used.

The fix is to include the templates in the package.

Fixes #5